Add spec for getting chapter list when response is correct, Fix test cases that break hystrix timeout config

+++ domain/chapter.go
@@ -1,10 +1,10 @@
 package domain
 
 type Chapter struct {
-	Number       int    `json:"hidden_chapter"`
-	Title        string `json:"judul"`
-	TitleId      string `json:"hidden_komik"`
-	ModifiedDate string `json:"waktu"`
+	Number       float32 `json:"hidden_chapter"`
+	Title        string  `json:"judul"`
+	TitleId      string  `json:"hidden_komik"`
+	ModifiedDate string  `json:"waktu"`
 }
 
 type ChapterListResponse struct {
